Refactor to make code more maintainable
